shell: I finally gave in and aliased g to git
many many people do this. I didn't do it before because I haven't
bothered to, and I didn't think 3-chara git is THAT much... But today I
was working with yadm (aliased to y) and then immediately switched to
git and undoubtedly 3-chars and 1-char is a huge difference

In any case, I don't think there are much (if not, any at all) cli
scripts with 1 letter names. I can't remember but I'm pretty sure that's
against the 'poetry of cli naming' article :P