Emacs setup script: Refactor, check version, add colors
- Check for emacs version before installing doom
- Check if emacs is installed before anything
- Add colors and boldness to errors and prompts
- Fixup indentation
- Fix invokation of doom at the end
- Consistency of output/messages

The way I did the colors was really messy and repetitive. I'm not sure
of the best solution for this yet. See, I want my scripts to be
standalone, which means they can be ran by themselves in a separate
environment so they shoudn't depend upon one another. Maybe I should
have like a dotscripts/lib/ansi to declare the color and text style
functions? idk