nvim: Fix Lazy's ft setting breaking ftplugin
The trick is to include 3 lines in init before lazy is loaded.

vim.cmd [[
  filetype plugin on
  filetype indent on
  syntax on
]]

Note that me switching ftdetect and ftplugin content to lua has nothing
to do with this fix. Vim should work well as well.

Made the switch to speed up operations. And fully eliminate Vimscript
from the config. (I love vimscript but unfortunately lua is faster!)